+++ frontend/src/components/itemComponents/form-add-item.tsx
@@ -4,7 +4,7 @@ import { z } from "zod";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import mongoose from "mongoose";
-import { postData } from "@/utils/functions";
+import { postData } from "@/utils/api";
 
 import {
   Form,
